move: 새 폴더 구조에 맞게 More과 Check의 하위페이지의 컴포넌트의 파일 이동
