+++ haul-fe/src/views/Check/CheckDetail/components/DriverInfoBox.jsx
@@ -1,8 +1,8 @@
 import styled from "styled-components";
-import Typography from "../../../components/Typhography/Typhography.jsx";
-import Margin from "../../../components/Margin/Margin.jsx";
-import UnderBar from "../../../components/UnderBar/UnderBar.jsx";
-import DummyDriverImg from "../../../assets/pngs/dummy.png";
+import Typography from "../../../../components/Typhography/Typhography.jsx";
+import Margin from "../../../../components/Margin/Margin.jsx";
+import UnderBar from "../../../../components/UnderBar/UnderBar.jsx";
+import DummyDriverImg from "../../../../assets/pngs/dummy.png";
 
 const DriverInfoDescription = {
   before: "차를 운전하실 기사님을 구하고 있어요.",
